requirements.txt
blinker==1.8.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.27.2
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
sympy==1.13.1
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
Werkzeug==3.0.3
tqdm